major push
major update to demotivator, including new insults, a new look, and a bit of new framework. The CSS now properly represents the style within the HTML. A few new insults were added, and we have a slight new look with a new font, which is way better then the old one. Be on the lookout for plenty of bug fixes later.

Cheers!
-PorkyProductions